+++ .auri/$6l1f7cok.md
@@ -1,0 +1,7 @@
+---
+package: "lucia-auth" # package name
+type: "minor" # "major", "minor", "patch"
+---
+
+`generateRandomString()` takes an optional second parameter - alphabet
+    - Remove `nanoid` dependency
