+++ .auri/$oyjxfoxw.md
@@ -1,0 +1,6 @@
+---
+package: "@lucia-auth/tokens" # package name
+type: "patch" # "major", "minor", "patch"
+---
+
+Remove `nanoid` dependency
