+++ examples/sveltekit-email/package.json
@@ -36,7 +36,6 @@
 		"@lucia-auth/adapter-prisma": "^1.0.0",
 		"@lucia-auth/tokens": "^1.0.0",
 		"@prisma/client": "^4.12.0",
-		"lucia-auth": "^1.0.0",
-		"nanoid": "^4.0.0"
+		"lucia-auth": "^1.0.0"
 	}
 }
