+++ packages/lucia-auth/package.json
@@ -46,7 +46,6 @@
 		"prettier": "^2.3.0"
 	},
 	"dependencies": {
-		"@noble/hashes": "^1.1.3",
-		"nanoid": "^4.0.0"
+		"@noble/hashes": "^1.1.3"
 	}
 }
