+++ test-apps/astro-email/src/auth/email.ts
@@ -1,5 +1,5 @@
 import { prismaClient } from "src/db";
-import { customAlphabet } from "nanoid";
+import { generateRandomString } from 'lucia-auth';
 import type { Email as DatabaseEmail } from "@prisma/client";
 
 const sendEmail = async (
@@ -7,12 +7,9 @@ const sendEmail = async (
 	subject: string,
 	content: string
 ) => {
-	const generateId = customAlphabet(
-		"0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
-	);
 	await prismaClient.email.create({
 		data: {
-			id: generateId(8),
+			id: generateRandomString(8),
 			subject,
 			email_address: emailAddress,
 			content,
